+++ tests/test_cli.py
@@ -1,9 +1,9 @@
 import json
+from typing import Dict, Any, List, Set
 
 import pytest
 
 from tests.testing_utils.fixtures import *
-from tests.testing_utils.result_assertions import find_conflicts_in_options_dict
 
 
 def test_global(runner):
@@ -37,3 +37,30 @@ def test_options_structure(runner):
 
     options_json = json.loads(result.output)
     assert find_conflicts_in_options_dict("snow", options_json) is None
+
+
+def find_conflicts_in_options_dict(path: str, options_dict: Dict[str, Any]):
+
+    keys = list(options_dict.keys())
+    duplicates = {}
+    if "options" in keys:
+        if opts_duplicates := check_options_for_duplicates(options_dict["options"]):
+            duplicates[path] = opts_duplicates
+        keys.remove("options")
+
+    for key in keys:
+        if key_duplicates := find_conflicts_in_options_dict(key, options_dict[key]):
+            duplicates[key] = key_duplicates
+
+    if duplicates:
+        return duplicates
+
+    return None
+
+
+def check_options_for_duplicates(options: Dict[str, List[str]]) -> Set[str]:
+    RESERVED_FLAGS = ["-h", "--help"]  # noqa: N806
+    flags = [flag for option in options.values() for flag in option]
+    return set(
+        [flag for flag in flags if (flags.count(flag) > 1 or flag in RESERVED_FLAGS)]
+    )
